Add serializers directory + refactor serializers #152
